assets: css: site: fix grid layout misconfiguration
This commit works in the sidebar using a CSS grid.

Step-by-step issue: When exiting the website listing pages
for a specific post, the sidebar was moving because the content
of the post was large and the responsiveness of the content was
not being handled correctly.

As there was a unit that collided with the static part of the
sidebar, it was possible to solve the problem using percentages
in columns-model-grid, as it is a responsive unit.

Signed-off-by: Aquila Macedo <aquila0dayprofissional@gmail.com>