assets: css: site: syling the contents
This commit works the structure and the way that
the content of the posts will be presented, a hierarchy
was added in the css using (CSS Inheritance, Cascade, and Specificity)
to better handle the content of the posts.

Overall this commit makes the site more stylish when presenting
the content. Using the pseudo-element "first-letter" in the
introduction of the text, a capital letter was implemented
for the beginning of the content and a new font was added for
the titles of the posts.

Signed-off-by: Aquila Macedo <aquila0dayprofissional@gmail.com>